Moved Datasets and Models to dedicated file. [WIP] Add optimizer.
